feat: update OpenTelemetry packages and enhance module architecture with discovery features
