+++ src/DevOpsMigrationPlatform.Abstractions/Services/IIdMapStoreFactory.cs
@@ -10,4 +10,10 @@ public interface IIdMapStoreFactory
     /// Creates a new <see cref="IIdMapStore"/> backed by the SQLite file at <paramref name="dbFilePath"/>.
     /// </summary>
     IIdMapStore Create(string dbFilePath);
+
+    /// <summary>
+    /// Creates a new <see cref="IIdMapStore"/> by resolving the idmap.db path from a
+    /// <c>file:///</c> package URI, including legacy path fallback.
+    /// </summary>
+    IIdMapStore CreateFromPackageUri(string packageUri);
 }
