+++ src/DevOpsMigrationPlatform.Infrastructure.AzureDevOps/DependencyServiceCollectionExtensions.cs
@@ -6,7 +6,6 @@ using DevOpsMigrationPlatform.Abstractions.Models;
 using DevOpsMigrationPlatform.Abstractions.Options;
 using DevOpsMigrationPlatform.Abstractions.Services;
 using DevOpsMigrationPlatform.Infrastructure.AzureDevOps.Factories;
-using DevOpsMigrationPlatform.Infrastructure.Modules;
 using DevOpsMigrationPlatform.Infrastructure.AzureDevOps.Services;
 using DevOpsMigrationPlatform.Infrastructure.Services;
 
@@ -88,7 +87,6 @@ public static class DependencyServiceCollectionExtensions
 
         // Factory for agent-side use where organisations come from DiscoveryJob (not config).
         services.AddSingleton<IDependencyDiscoveryServiceFactory, DependencyDiscoveryServiceFactory>();
-        services.AddSingleton<IDiscoveryModule, DependencyDiscoveryModule>();
 
         return services;
     }
